Brings in changes that remove Easel's augmentation concept, and make
build robust against gnu autoheader.

+++ src/hmmc2.c
@@ -176,7 +176,7 @@ usage(char *pgm)
 int main(int argc, char *argv[])
 {
   int              i, j;
-  int              n;
+  uint64_t              n;
   int              eod;
   int              size;
